Refactored the frontend The frontend now resides in its own directory The configuration files to the frontend are now grouped under the config directory The static and template folders have been moved into the frontend directory Separated the authentication and API endpoints Separated the frontend into a core class containing the Flask application and its subcomponents Created a config class for the frontend to manage configuration Created a factory class to manufacture endpoint routes
