--- ga4gh/server/config/oidc_auth_config.yml
@@ -1,30 +1,0 @@
-frontend:
-    OIDC_CLIENT_SECRETS: /srv/ga4gh-server/client_secrets.json
-    OIDC_OPENID_REALM: http://192.168.99.100:8000/oidc_callback
-    OIDC_ID_TOKEN_COOKIE_SECURE: False
-    OIDC_REQUIRE_VERIFIED_EMAIL: False
-    SECRET_KEY: key
-    TESTING: False
-    DEBUG: False
-server:
-    MAX_CONTENT_LENGTH: 2097152
-    MAX_RESPONSE_LENGTH: 1048576 
-    REQUEST_VALIDATION: True
-    DEFAULT_PAGE_SIZE: 100
-    FILE_HANDLE_CACHE_MAX_SIZE: 50
-    LANDING_MESSAGE_HTML: landing_message.html
-    DATA_SOURCE: /srv/ga4gh-compliance-data/registry.db
-    INITIAL_PEERS: /srv/ga4gh-server/ga4gh/server/templates/initial_peers.txt
-    DEBUG: True
-    SIMULATED_BACKEND_RANDOM_SEED: 0
-    SIMULATED_BACKEND_NUM_CALLS: 1
-    SIMULATED_BACKEND_VARIANT_DENSITY: 0.5
-    SIMULATED_BACKEND_NUM_VARIANT_SETS: 1
-    SIMULATED_BACKEND_NUM_REFERENCE_SETS: 1
-    SIMULATED_BACKEND_NUM_REFERENCES_PER_REFERENCE_SET: 1
-    SIMULATED_BACKEND_NUM_ALIGNMENTS_PER_READ_GROUP: 2
-    SIMULATED_BACKEND_NUM_READ_GROUPS_PER_READ_GROUP_SET: 2
-    SIMULATED_BACKEND_NUM_PHENOTYPE_ASSOCIATIONS: 2
-    SIMULATED_BACKEND_NUM_PHENOTYPE_ASSOCIATION_SETS: 2
-    SIMULATED_BACKEND_NUM_RNA_QUANTIFICATION_SETS: 2
-    SIMULATED_BACKEND_NUM_EXPRESSION_LEVELS_PER_RNA_QUANT_SET: 2
