+++ ga4gh/server/frontend/config/full_config.yml
@@ -1,0 +1,63 @@
+MAX_CONTENT_LENGTH: 2097152
+MAX_RESPONSE_LENGTH: 1048576 
+REQUEST_VALIDATION: True
+DEFAULT_PAGE_SIZE: 100
+FILE_HANDLE_CACHE_MAX_SIZE: 50
+LANDING_MESSAGE_HTML: landing_message.html
+DATA_SOURCE: /srv/ga4gh-compliance-data/registry.db
+INITIAL_PEERS: /srv/ga4gh-server/ga4gh/server/templates/initial_peers.txt
+CACHE_DIRECTORY: /tmp/ga4gh
+DEBUG: True
+TESTING: False
+KEYCLOAK: True
+SECRET_KEY: super_secret
+APPLICATION_ROOT: None
+JSONIFY_PRETTYPRINT_REGULAR: True
+JSON_AS_ASCII: True
+JSON_SORT_KEYS: True
+LOGGER_NAME: ga4gh.frontend
+MAX_CONTENT_LENGTH: None
+PERMANENT_SESSION_LIFETIME: 31 days, 0:00:00
+PREFERRED_URL_SCHEME: http
+PRESERVE_CONTEXT_ON_EXCEPTION: None
+PROPAGATE_EXCEPTIONS: None
+SEND_FILE_MAX_AGE_DEFAULT: 43200
+SERVER_NAME: None
+SESSION_COOKIE_DOMAIN: None
+SESSION_COOKIE_HTTPONLY: True
+SESSION_COOKIE_NAME: session
+SESSION_COOKIE_PATH: None
+SESSION_COOKIE_SECURE: False
+TRAP_BAD_REQUEST_ERRORS: False
+TRAP_HTTP_EXCEPTIONS: False
+USE_X_SENDFILE: False
+OIDC_CLIENT_SECRETS: /srv/ga4gh-server/client_secrets.json
+OIDC_OPENID_REALM: http://192.168.99.100:8000/oidc_callback
+OIDC_ID_TOKEN_COOKIE_SECURE: False
+OIDC_REQUIRE_VERIFIED_EMAIL: False
+OIDC_PROVIDER: https://opennet-33-237.uhnres.utoronto.ca:8443
+OIDC_REDIRECT_URI: https://localhost/oauth2callback
+OIDC_CLIENT_ID: XXX
+OIDC_CLIENT_SECRET: XXX
+OIDC_AUTHZ_ENDPOINT: http://localhost:8080/auth/realms/demo/protocol/openid-connect/auth   
+OIDC_TOKEN_ENDPOINT: http://localhost:8080/auth/realms/demo/protocol/openid-connect/token
+OIDC_TOKEN_REV_ENDPOINT: http://localhost:8080/auth/realms/"demo/protocol/openid-connect/token/introspect
+AUTH0_ENABLED: False
+AUTH0_SCOPES: openid email
+AUTH0_CALLBACK_URL: http://localhost:8000/callback
+AUTH0_HOST: david4096.auth0.com
+AUTH0_CLIENT_ID: r99hdj5hhkazgePB5oMYK9Sv4NaUwwYp
+AUTH0_CLIENT_SECRET: KeV2tMyGaSgLeOhpoGs_XLH65Tfw43yBjT8DIpaTxXAKmd_bguJwXA6T7D0iYfgB
+AUTH0_AUTHORIZED_EMAILS: davidcs@ucsc.edu,your@email.com
+SIMULATED_BACKEND_RANDOM_SEED: 0
+SIMULATED_BACKEND_NUM_CALLS: 1
+SIMULATED_BACKEND_VARIANT_DENSITY: 0.5
+SIMULATED_BACKEND_NUM_VARIANT_SETS: 1
+SIMULATED_BACKEND_NUM_REFERENCE_SETS: 1
+SIMULATED_BACKEND_NUM_REFERENCES_PER_REFERENCE_SET: 1
+SIMULATED_BACKEND_NUM_ALIGNMENTS_PER_READ_GROUP: 2
+SIMULATED_BACKEND_NUM_READ_GROUPS_PER_READ_GROUP_SET: 2
+SIMULATED_BACKEND_NUM_PHENOTYPE_ASSOCIATIONS: 2
+SIMULATED_BACKEND_NUM_PHENOTYPE_ASSOCIATION_SETS: 2
+SIMULATED_BACKEND_NUM_RNA_QUANTIFICATION_SETS: 2
+SIMULATED_BACKEND_NUM_EXPRESSION_LEVELS_PER_RNA_QUANT_SET: 2
