--- ga4gh/server/frontend/frontend.py
@@ -1,1170 +1,0 @@
-"""
-The Flask frontend for the GA4GH API.
-
-TODO Document properly.
-"""
-
-from __future__ import division
-from __future__ import print_function
-from __future__ import unicode_literals
-
-import os
-import socket
-import urlparse
-import functools
-import json
-
-import flask
-import flask.ext.cors as cors
-from flask import request
-from flask.ext.oidc import OpenIDConnect
-import werkzeug
-import oic
-import oic.oauth2
-import oic.oic.message as message
-from oauth2client.client import OAuth2Credentials
-import requests
-import logging
-from logging import StreamHandler
-from werkzeug.contrib.cache import FileSystemCache
-from yaml import load
-import pkg_resources
-
-import ga4gh.server.backend as backend
-import ga4gh.server.datamodel as datamodel
-import ga4gh.server.exceptions as exceptions
-import ga4gh.server.datarepo as datarepo
-import ga4gh.server.auth as auth
-import ga4gh.server.network as network
-import ga4gh.schemas.protocol as protocol
-import ga4gh.server.frontend.flaskStatus as flaskStatus
-
-MIMETYPE = "application/json"
-SEARCH_ENDPOINT_METHODS = ['POST', 'OPTIONS']
-SECRET_KEY_LENGTH = 24
-
-
-def import_yaml_config(config):
-    """
-    A function that returns the frontend and server configuration
-    as a dictionary.
-    """
-    stream = file(config, 'r')
-    config = load(stream)
-    return config
-
-
-app = flask.Flask(__name__)
-
-assert not hasattr(app, 'urls')
-app.urls = []
-requires_auth = auth.auth_decorator(app)
-
-# Edit this for flask-oidc, the endpoints are in the client_secrets.json file
-pathLocation = '/'.join(('..', 'config', 'oidc_auth_config.yml'))
-configPath = pkg_resources.resource_filename(__name__, pathLocation)
-config = import_yaml_config(config=configPath)
-app.config.update(config["frontend"])
-
-# For configuration of Flask-Oidc
-oidc = OpenIDConnect(app)
-
-
-class NoConverter(werkzeug.routing.BaseConverter):
-    """
-    A converter that allows the routing matching algorithm to not
-    match on certain literal terms
-
-    This is needed because if there are e.g. two routes:
-
-    /callsets/search
-    /callsets/<id>
-
-    A request for /callsets/search will get routed to
-    the second, which is not what we want.
-    """
-    def __init__(self, map, *items):
-        werkzeug.routing.BaseConverter.__init__(self, map)
-        self.items = items
-
-    def to_python(self, value):
-        if value in self.items:
-            raise werkzeug.routing.ValidationError()
-        return value
-
-
-app.url_map.converters['no'] = NoConverter
-
-
-
-def reset():
-    """
-    Resets the flask app; used in testing
-    """
-    app.config.clear()
-    configStr = 'ga4gh.server.serverconfig:FlaskDefaultConfig'
-    app.config.from_object(configStr)
-
-
-def _configure_backend(app):
-    """A helper function used just to help modularize the code a bit."""
-    # Allocate the backend
-    # We use URLs to specify the backend. Currently we have file:// URLs (or
-    # URLs with no scheme) for the SqlDataRepository, and special empty:// and
-    # simulated:// URLs for empty or simulated data sources.
-    dataSource = urlparse.urlparse(app.config["DATA_SOURCE"], "file")
-
-    if dataSource.scheme == "simulated":
-        # Ignore the query string
-        randomSeed = app.config["SIMULATED_BACKEND_RANDOM_SEED"]
-        numCalls = app.config["SIMULATED_BACKEND_NUM_CALLS"]
-        variantDensity = app.config["SIMULATED_BACKEND_VARIANT_DENSITY"]
-        numVariantSets = app.config["SIMULATED_BACKEND_NUM_VARIANT_SETS"]
-        numReferenceSets = app.config[
-            "SIMULATED_BACKEND_NUM_REFERENCE_SETS"]
-        numReferencesPerReferenceSet = app.config[
-            "SIMULATED_BACKEND_NUM_REFERENCES_PER_REFERENCE_SET"]
-        numAlignmentsPerReadGroup = app.config[
-            "SIMULATED_BACKEND_NUM_ALIGNMENTS_PER_READ_GROUP"]
-        numReadGroupsPerReadGroupSet = app.config[
-            "SIMULATED_BACKEND_NUM_READ_GROUPS_PER_READ_GROUP_SET"]
-        numPhenotypeAssociations = app.config[
-            "SIMULATED_BACKEND_NUM_PHENOTYPE_ASSOCIATIONS"]
-        numPhenotypeAssociationSets = app.config[
-            "SIMULATED_BACKEND_NUM_PHENOTYPE_ASSOCIATION_SETS"]
-        numRnaQuantSets = app.config[
-            "SIMULATED_BACKEND_NUM_RNA_QUANTIFICATION_SETS"]
-        numExpressionLevels = app.config[
-            "SIMULATED_BACKEND_NUM_EXPRESSION_LEVELS_PER_RNA_QUANT_SET"]
-
-        dataRepository = datarepo.SimulatedDataRepository(
-            randomSeed=randomSeed, numCalls=numCalls,
-            variantDensity=variantDensity, numVariantSets=numVariantSets,
-            numReferenceSets=numReferenceSets,
-            numReferencesPerReferenceSet=numReferencesPerReferenceSet,
-            numReadGroupsPerReadGroupSet=numReadGroupsPerReadGroupSet,
-            numAlignments=numAlignmentsPerReadGroup,
-            numPhenotypeAssociations=numPhenotypeAssociations,
-            numPhenotypeAssociationSets=numPhenotypeAssociationSets,
-            numRnaQuantSets=numRnaQuantSets,
-            numExpressionLevels=numExpressionLevels)
-    elif dataSource.scheme == "empty":
-        dataRepository = datarepo.EmptyDataRepository()
-    elif dataSource.scheme == "file":
-        path = os.path.join(dataSource.netloc, dataSource.path)
-        dataRepository = datarepo.SqlDataRepository(path)
-        dataRepository.open(datarepo.MODE_READ)
-    else:
-        raise exceptions.ConfigurationException(
-            "Unsupported data source scheme: " + dataSource.scheme)
-    theBackend = backend.Backend(dataRepository)
-    theBackend.setRequestValidation(app.config["REQUEST_VALIDATION"])
-    theBackend.setDefaultPageSize(app.config["DEFAULT_PAGE_SIZE"])
-    theBackend.setMaxResponseLength(app.config["MAX_RESPONSE_LENGTH"])
-    return theBackend
-
-
-def configure(configFile=None, baseConfig="ProductionConfig",
-              port=8000, extraConfig={}):
-    """
-    TODO Document this critical function! What does it do?
-    What does it assume?
-
-    Based on the configuration the server is being hosted with,
-    it initalizes all the variables needed, and generates the
-    redirect-url for the Auth and OIDC providers (if present).
-
-    """
-    file_handler = StreamHandler()
-    file_handler.setLevel(logging.WARNING)
-    app.logger.addHandler(file_handler)
-    configStr = 'ga4gh.server.serverconfig:{0}'.format(baseConfig)
-    app.config.from_object(configStr)
-    if os.environ.get('GA4GH_CONFIGURATION') is not None:
-        app.config.from_envvar('GA4GH_CONFIGURATION')
-    if configFile is not None:
-        app.config.from_pyfile(configFile)
-    app.config.update(extraConfig.items())
-    # Setup file handle cache max size
-    datamodel.fileHandleCache.setMaxCacheSize(
-        app.config["FILE_HANDLE_CACHE_MAX_SIZE"])
-    # Setup CORS
-    try:
-        cors.CORS(app, allow_headers='Content-Type')
-    except AssertionError:
-        pass
-    app.serverStatus = flaskStatus.ServerStatus(app)
-
-    app.backend = _configure_backend(app)
-
-    if app.config.get('SECRET_KEY'):
-        app.secret_key = app.config['SECRET_KEY']
-    elif app.config.get('OIDC_PROVIDER'):
-        raise exceptions.ConfigurationException(
-            'OIDC configuration requires a secret key')
-    if app.config.get('CACHE_DIRECTORY'):
-        app.cache_dir = app.config['CACHE_DIRECTORY']
-    else:
-        app.cache_dir = '/tmp/ga4gh'
-    app.cache = FileSystemCache(
-        app.cache_dir, threshold=5000, default_timeout=600, mode=384)
-    # Peer service initialization
-    network.initialize(
-        app.config.get('INITIAL_PEERS'),
-        app.backend.getDataRepository(),
-        app.logger)
-    app.oidcClient = None
-    app.myPort = port
-    if app.config.get('AUTH0_ENABLED'):
-        emails = app.config.get('AUTH0_AUTHORIZED_EMAILS', '').split(',')
-        [auth.authorize_email(e, app.cache) for e in emails]
-    if "OIDC_PROVIDER" in app.config:
-        # The oic client. If we're testing, we don't want to verify
-        # SSL certificates
-        app.oidcClient = oic.oic.Client(
-            verify_ssl=('TESTING' not in app.config))
-        try:
-            app.oidcClient.provider_config(app.config['OIDC_PROVIDER'])
-        except requests.exceptions.ConnectionError:
-            configResponse = message.ProviderConfigurationResponse(
-                issuer=app.config['OIDC_PROVIDER'],
-                authorization_endpoint=app.config['OIDC_AUTHZ_ENDPOINT'],
-                token_endpoint=app.config['OIDC_TOKEN_ENDPOINT'],
-                revocation_endpoint=app.config['OIDC_TOKEN_REV_ENDPOINT'])
-            app.oidcClient.handle_provider_config(configResponse,
-                                                  app.config['OIDC_PROVIDER'])
-
-        # The redirect URI comes from the configuration.
-        # If we are testing, then we allow the automatic creation of a
-        # redirect uri if none is configured
-        redirectUri = app.config.get('OIDC_REDIRECT_URI')
-        if redirectUri is None:
-            redirectUri = 'https://{0}:{1}/oauth2callback'.format(
-                socket.gethostname(), app.myPort)
-        app.oidcClient.redirect_uris = [redirectUri]
-        if redirectUri is []:
-            raise exceptions.ConfigurationException(
-                'OIDC configuration requires a redirect uri')
-        print(redirectUri)
-        # We only support dynamic registration while testing.
-        if ('registration_endpoint' in app.oidcClient.provider_info):
-            app.oidcClient.register(
-                app.oidcClient.provider_info["registration_endpoint"],
-                redirect_uris=[redirectUri])
-        else:
-            response = message.RegistrationResponse(
-                client_id=app.config['OIDC_CLIENT_ID'],
-                client_secret=app.config['OIDC_CLIENT_SECRET'],
-                redirect_uris=[redirectUri],
-                verify_ssl=False)
-            app.oidcClient.store_registration_info(response)
-    # This is for the configuration of the Keycloak Server
-    # added by Kevin Chan on June 12, 2017
-    """
-    if "KEYCLOAK" in app.config:
-        # Configuration using the requests library.
-        # I left this here for your reference to
-        # familiarize yourself with the Authorization code flow
-        # Makes a request to get the configuration,
-        # and stores all the endpoints/id/secrets needed
-
-        app.oidcClient = Client(client_authn_method=CLIENT_AUTHN_METHOD)
-        endpoints = requests.get(app.config["WELL_KNOWN_CONFIG"]).json()
-        opInfo = message.ProviderConfigurationResponse(
-            issuer=endpoints["issuer"],
-            authorization_endpoint=endpoints["authorization_endpoint"],
-            token_endpoint=endpoints["token_endpoint"],
-            introspect_endpoint=endpoints["token_introspection_endpoint"],
-            user_endpoint= endpoints["userinfo_endpoint"],
-            session_endpoint=endpoints["end_session_endpoint"],
-            jwks_uri= endpoints["jwks_uri"])
-        app.oidcClient.provider_info = opInfo
-        app.oidcClient.client_id = app.config['CLIENT_ID']
-        app.oidcClient.client_secret = app.config["CLIENT_SECRET"]
-
-        #For configuration of your server, change the host.
-        redirectUri = 'http://{0}:{1}/keycallback'.format(
-                socket.gethostbyname(socket.gethostname()), app.myPort)
-        app.oidcClient.redirect_uris = [redirectUri]
-    """
-
-
-def getFlaskResponse(responseString, httpStatus=200):
-    """
-    Returns a Flask response object for the specified data and HTTP status.
-    """
-    return flask.Response(responseString, status=httpStatus, mimetype=MIMETYPE)
-
-
-def handleHttpPost(request, endpoint):
-    """
-    Handles the specified HTTP POST request, which maps to the specified
-    protocol handler endpoint and protocol request class.
-    """
-    if request.mimetype and request.mimetype != MIMETYPE:
-        raise exceptions.UnsupportedMediaTypeException()
-    request = request.get_data()
-    if request == '' or request is None:
-        request = '{}'
-    responseStr = endpoint(request)
-    return getFlaskResponse(responseStr)
-
-
-def handleList(endpoint, request):
-    """
-    Handles the specified HTTP GET request, mapping to a list request
-    """
-    responseStr = endpoint(request.get_data())
-    return getFlaskResponse(responseStr)
-
-
-def handleHttpGet(id_, endpoint):
-    """
-    Handles the specified HTTP GET request, which maps to the specified
-    protocol handler endpoint and protocol request class
-    """
-    responseStr = endpoint(id_)
-    return getFlaskResponse(responseStr)
-
-
-def handleHttpOptions():
-    """
-    Handles the specified HTTP OPTIONS request.
-    """
-    response = getFlaskResponse("")
-    response.headers.add("Access-Control-Request-Methods", "GET,POST,OPTIONS")
-    return response
-
-
-@app.errorhandler(Exception)
-def handleException(exception):
-    """
-    Handles an exception that occurs somewhere in the process of handling
-    a request.
-    """
-    serverException = exception
-    if not isinstance(exception, exceptions.BaseServerException):
-        with app.test_request_context():
-            app.log_exception(exception)
-        serverException = exceptions.getServerError(exception)
-    error = serverException.toProtocolElement()
-    # If the exception is being viewed by a web browser, we can render a nicer
-    # view.
-    if flask.request and 'Accept' in flask.request.headers and \
-            flask.request.headers['Accept'].find('text/html') != -1:
-        message = "<h1>Error {}</h1><pre>{}</pre>".format(
-                    serverException.httpStatus,
-                    protocol.toJson(error))
-        if serverException.httpStatus == 401 \
-                or serverException.httpStatus == 403:
-            message += "Please try <a href=\"/login\">logging in</a>."
-        return message
-    else:
-        responseStr = protocol.toJson(error)
-        return getFlaskResponse(responseStr, serverException.httpStatus)
-
-
-# Added by Kevin Chan
-def requires_token(f):
-    """
-    Decorator function that ensures that the token is valid,
-    if the token is invalid or expired, the user will be
-    redirected to the login page. Much of the authorization
-    code flow is done solely by the function decorator
-    @oidc.require_login
-    """
-    @functools.wraps(f)
-    def decorated(*args, **kargs):
-        if app.config.get("KEYCLOAK"):
-            hostname = socket.gethostbyname(socket.gethostname())
-            redirectUri = 'http://{0}:{1}{2}'.format(
-                    hostname, app.myPort, request.path)
-            try:
-                info = oidc.user_getinfo(['sub'])
-                subStore = oidc.credentials_store[info.get('sub')]
-                credentials = OAuth2Credentials.from_json(subStore)
-                tokenResponse = credentials.token_response
-                introspectArgs = {
-                    "token": tokenResponse["access_token"],
-                    "client_id": oidc.client_secrets["client_id"],
-                    "client_secret": oidc.client_secrets["client_secret"],
-                    "refresh_token": tokenResponse["refresh_token"],
-                }
-            except:
-                return flask.redirect(redirectUri)
-            introspect = oidc.client_secrets["token_introspection_uri"]
-            userInfo = requests.post(url=introspect, data=introspectArgs)
-
-            if userInfo.status_code != 200:
-                raise exceptions.NotAuthenticatedException()
-        return f(*args, **kargs)
-    return decorated
-
-
-def startLogin():
-    """
-    If user is not logged in, this generates the redirect URL
-    to the OIDC or Auth provider (depending on the configuration)
-    Returns: the redirect response
-    """
-    flask.session["state"] = oic.oauth2.rndstr(SECRET_KEY_LENGTH)
-    flask.session["nonce"] = oic.oauth2.rndstr(SECRET_KEY_LENGTH)
-    provInfo = app.oidcClient.provider_info
-    args = {
-        "client_id": app.oidcClient.client_id,
-        "response_type": "code",
-        "scope": ["openid", "profile"],
-        "nonce": flask.session["nonce"],
-        "redirect_uri": app.oidcClient.redirect_uris[0],
-        "authorization_endpoint": provInfo["authorization_endpoint"],
-        "state": flask.session["state"],
-    }
-
-    # First condition is the configuration for the Keycloak Server.
-    # Redirects the user to the Keycloak sign in page.
-    # I left this here for your reference.
-    # Added by Kevin Chan
-
-    if "WELL_KNOWN_CONFIG" in app.config:
-        # result = app.oidcClient.do_authorization_request(
-        #     request_args=args, state=flask.session["state"])
-        result = app.oidcClient.construct_AuthorizationRequest(
-            request_args=args)
-        addOn = result.request(app.oidcClient.authorization_endpoint)
-        authInfo = app.oidcClient.provider_info["authorization_endpoint"]
-        loginUrl = authInfo + addOn
-
-        if request.path == "/login":
-            flask.session["path"] = "/"
-        else:
-            flask.session["path"] = request.path
-        return flask.redirect(loginUrl)
-
-    result = app.oidcClient.do_authorization_request(
-        request_args=args, state=flask.session["state"])
-    return flask.redirect(result.url)
-
-
-# Commented out by Kevin Chan.
-# If Using the Keycloak Config or no authentication
-# this function is no longer needed
-"""
-@app.before_request
-def checkAuthentication():
-
-    The request will have a parameter 'key' if it came from the command line
-    client, or have a session key of 'key' if it's the browser.
-    If the token is not found, start the login process.
-
-    If there is no oidcClient, we are runnning naked and we don't check.
-    If we're being redirected to the oidcCallback we don't check.
-
-    :returns None if all is ok (and the request handler continues as usual).
-    Otherwise if the key was in the session (therefore we're in a browser)
-    then startLogin() will redirect to the OIDC provider. If the key was in
-    the request arguments, we're using the command line and just raise an
-    exception.
-
-    if app.oidcClient is None:
-        return
-
-    oidcCall = flask.request.endpoint == 'oidcCallback'
-    keyCall = flask.request.endpoint == "keycloakCallback"
-    if oidcCall or keyCall:
-        return
-    key = flask.session.get('key') or flask.request.args.get('key')
-
-    if key is None: # or not app.cache.get(key):
-        if 'key' in flask.request.args:
-            raise exceptions.NotAuthorizedException()
-        else:
-            return startLogin()
-"""
-
-
-def handleFlaskGetRequest(id_, flaskRequest, endpoint):
-    """
-    Handles the specified flask request for one of the GET URLs
-    Invokes the specified endpoint to generate a response.
-    """
-    if flaskRequest.method == "GET":
-        return handleHttpGet(id_, endpoint)
-    else:
-        raise exceptions.MethodNotAllowedException()
-
-
-def handleFlaskListRequest(id_, flaskRequest, endpoint):
-    """
-    Handles the specified flask list request for one of the GET URLs.
-    Invokes the specified endpoint to generate a response.
-    """
-
-    return handleList(endpoint, flaskRequest)
-
-
-def handleFlaskPostRequest(flaskRequest, endpoint):
-    """
-    Handles the specified flask request for one of the POST URLS
-    Invokes the specified endpoint to generate a response.
-    """
-    if flaskRequest.method == "POST":
-        return handleHttpPost(flaskRequest, endpoint)
-    elif flaskRequest.method == "OPTIONS":
-        return handleHttpOptions()
-    else:
-        raise exceptions.MethodNotAllowedException()
-
-
-class DisplayedRoute(object):
-    """
-    Registers that a route should be displayed on the html page
-    """
-    def __init__(
-            self, path, postMethod=False, pathDisplay=None):
-        self.path = path
-        self.methods = None
-        if postMethod:
-            methodDisplay = 'POST'
-            self.methods = SEARCH_ENDPOINT_METHODS
-        else:
-            methodDisplay = 'GET'
-        if pathDisplay is None:
-            pathDisplay = path
-        app.urls.append((methodDisplay, pathDisplay))
-
-    def __call__(self, func):
-        if self.methods is None:
-            app.add_url_rule(self.path, func.func_name, func)
-        else:
-            app.add_url_rule(
-                self.path, func.func_name, func, methods=self.methods)
-
-        @functools.wraps(func)
-        def wrapper(*args, **kwargs):
-            result = func(*args, **kwargs)
-            return result
-        return wrapper
-
-
-@app.route('/')
-@oidc.require_login
-@requires_token
-def index():
-    response = flask.render_template('index.html',
-                                     info=app.serverStatus)
-    if app.config.get('AUTH0_ENABLED'):
-        key = (flask.request.args.get('key'))
-        try:
-            print(key)
-            profile = app.cache.get(key)
-        except:
-            raise exceptions.NotAuthorizedException()
-        if (profile):
-            return response
-        else:
-            exceptions.NotAuthenticatedException()
-    else:
-        return response
-
-
-# New configuration added by Kevin Chan
-@app.route("/login")
-def login():
-
-    if app.config.get('AUTH0_ENABLED'):
-        if (flask.request.args.get('code')):
-            return auth.render_key(app, key=flask.request.args.get('code'))
-        else:
-            return auth.render_login(
-                app=app,
-                scopes=app.config.get('AUTH0_SCOPES'),
-                redirect_uri=app.config.get('AUTH0_CALLBACK_URL'),
-                domain=app.config.get('AUTH0_HOST'),
-                client_id=app.config.get('AUTH0_CLIENT_ID'))
-
-    # Configuration for KeyCloak Server
-    elif app.config.get('KEYCLOAK'):
-        app.oidClient = None
-        flask.session.clear()
-        return startLogin()
-
-    else:
-        raise exceptions.NotFoundException()
-
-
-@app.route('/callback')
-def callback_handling():
-    if app.config.get('AUTH0_ENABLED'):
-        return auth.callback_maker(
-            cache=app.cache,
-            domain=app.config.get('AUTH0_HOST'),
-            client_id=app.config.get('AUTH0_CLIENT_ID'),
-            client_secret=app.config.get('AUTH0_CLIENT_SECRET'),
-            redirect_uri=app.config.get('AUTH0_CALLBACK_URL'))()
-    else:
-        raise exceptions.NotFoundException()
-
-
-@app.route("/logout")
-@requires_auth
-@cors.cross_origin(headers=['Content-Type', 'Authorization'])
-def logout():
-    if app.config.get('AUTH0_ENABLED'):
-        key = flask.session['auth0_key']
-        auth.logout(app.cache)
-        url = 'https://{}/v2/logout?access_token={}&?client_id={}'.format(
-            app.config.get('AUTH0_HOST'),
-            key,
-            app.config.get('AUTH0_CLIENT_ID'),
-            app.config.get('AUTH0_CALLBACK_URL'))
-        return flask.redirect(url)
-    else:
-
-        # args = {
-        #    "token": app.oidcClient.token["access_token"],
-        #    "client_id": app.oidcClient.client_id,
-        #    "client_secret": app.oidcClient.client_secret,
-        #    "refresh_token": app.oidcClient.token["refresh_token"]
-        #    }
-
-        # endpoint = app.oidcClient.provider_info["session_endpoint"]
-        # logout = requests.post(url=endpoint, data=args)
-        flask.session.clear()
-        hostname = socket.gethostbyname(socket.gethostname())
-        return flask.redirect("http://{0}:{1}".format(hostname, app.myPort))
-
-
-@app.route('/favicon.ico')
-@app.route('/robots.txt')
-def robots():
-    return flask.send_from_directory(
-        app.static_folder, flask.request.path[1:])
-
-
-@DisplayedRoute('/info')
-@requires_auth
-@oidc.require_login
-def getInfo():
-    action = app.backend.runGetInfo
-    return handleFlaskGetRequest(None, flask.request, action)
-
-
-@DisplayedRoute('/references/<id>')
-@requires_auth
-@oidc.require_login
-@requires_token
-def getReference(id):
-    action = app.backend.runGetReference
-    return handleFlaskGetRequest(id, flask.request, action)
-
-
-@DisplayedRoute('/referencesets/<id>')
-@requires_auth
-@oidc.require_login
-@requires_token
-def getReferenceSet(id):
-    action = app.backend.runGetReferenceSet
-    return handleFlaskGetRequest(id, flask.request, action)
-
-
-@DisplayedRoute('/listreferencebases', postMethod=True)
-def listReferenceBases():
-    action = app.backend.runListReferenceBases
-    return handleFlaskListRequest(id, flask.request, action)
-
-
-@DisplayedRoute('/callsets/search', postMethod=True)
-def searchCallSets():
-    action = app.backend.runSearchCallSets
-    return handleFlaskPostRequest(flask.request, action)
-
-
-@DisplayedRoute('/readgroupsets/search', postMethod=True)
-def searchReadGroupSets():
-    action = app.backend.runSearchReadGroupSets
-    return handleFlaskPostRequest(flask.request, action)
-
-
-@DisplayedRoute('/reads/search', postMethod=True)
-def searchReads():
-    return handleFlaskPostRequest(
-        flask.request, app.backend.runSearchReads)
-
-
-@DisplayedRoute('/referencesets/search', postMethod=True)
-def searchReferenceSets():
-    action = app.backend.runSearchReferenceSets
-    return handleFlaskPostRequest(flask.request, action)
-
-
-@DisplayedRoute('/references/search', postMethod=True)
-def searchReferences():
-    action = app.backend.runSearchReferences
-    return handleFlaskPostRequest(
-        flask.request, action)
-
-
-@DisplayedRoute('/variantsets/search', postMethod=True)
-def searchVariantSets():
-    action = app.backend.runSearchVariantSets
-    return handleFlaskPostRequest(flask.request, action)
-
-
-@DisplayedRoute('/variants/search', postMethod=True)
-def searchVariants():
-    action = app.backend.runSearchVariant
-    return handleFlaskPostRequest(flask.request, action)
-
-
-@DisplayedRoute('/variantannotationsets/search', postMethod=True)
-def searchVariantAnnotationSets():
-    action = app.backend.runSearchVariantAnnotationSets
-    return handleFlaskPostRequest(flask.request, action)
-
-
-@DisplayedRoute('/variantannotations/search', postMethod=True)
-def searchVariantAnnotations():
-    action = app.backend.runSearchVariantAnnotations
-    return handleFlaskPostRequest(flask.request, action)
-
-
-@DisplayedRoute('/datasets/search', postMethod=True)
-@requires_auth
-@oidc.require_login
-@requires_token
-def searchDatasets():
-    action = app.backend.runSearchDatasets
-    return handleFlaskPostRequest(flask.request, action)
-
-
-@DisplayedRoute('/featuresets/search', postMethod=True)
-@requires_auth
-@oidc.require_login
-@requires_token
-def searchFeatureSets():
-    action = app.backend.runSearchFeatureSets
-    return handleFlaskPostRequest(flask.request, action)
-
-
-@DisplayedRoute('/features/search', postMethod=True)
-@requires_auth
-@oidc.require_login
-@requires_token
-def searchFeatures():
-    action = app.backend.runSearchFeatures
-    return handleFlaskPostRequest(flask.request, action)
-
-
-@DisplayedRoute('/continuoussets/search', postMethod=True)
-@requires_auth
-@oidc.require_login
-@requires_token
-def searchContinuousSets():
-    action = app.backend.runSearchContinuousSets
-    return handleFlaskPostRequest(flask.request, action)
-
-
-@DisplayedRoute('/continuous/search', postMethod=True)
-@requires_auth
-@oidc.require_login
-@requires_token
-def searchContinuous():
-    action = app.backend.runSearchContinuous
-    return handleFlaskPostRequest(flask.request, action)
-
-
-@DisplayedRoute('/biosamples/search', postMethod=True)
-@requires_auth
-@oidc.require_login
-@requires_token
-def searchBiosamples():
-    action = app.backend.runSearchBiosamples
-    return handleFlaskPostRequest(flask.request, action)
-
-
-@DisplayedRoute('/individuals/search', postMethod=True)
-@requires_auth
-@oidc.require_login
-@requires_token
-def searchIndividuals():
-    action = app.backend.runSearchIndividuals
-    return handleFlaskPostRequest(flask.request, action)
-
-
-@DisplayedRoute('/peers/list', postMethod=True)
-@requires_auth
-@oidc.require_login
-@requires_token
-def listPeers():
-    action = app.backend.runListPeers
-    return handleFlaskPostRequest(flask.request, action)
-
-
-@DisplayedRoute('/announce', postMethod=True)
-@requires_auth
-@oidc.require_login
-@requires_token
-def announce():
-    # We can't use the post handler here because we want detailed request
-    # data.
-    return app.backend.runAddAnnouncement(flask.request)
-
-
-@DisplayedRoute(
-    '/biosamples/<no(search):id>',
-    pathDisplay='/biosamples/<id>')
-@requires_auth
-@oidc.require_login
-@requires_token
-def getBiosample(id):
-    action = app.backend.runGetBiosample
-    return handleFlaskGetRequest(id, flask.request, action)
-
-
-@DisplayedRoute(
-    '/individuals/<no(search):id>',
-    pathDisplay='/individuals/<id>')
-@requires_auth
-@oidc.require_login
-@requires_token
-def getIndividual(id):
-    action = app.backend.runGetIndividual
-    return handleFlaskGetRequest(id, flask.request, action)
-
-
-@DisplayedRoute('/rnaquantificationsets/search', postMethod=True)
-@requires_auth
-@oidc.require_login
-@requires_token
-def searchRnaQuantificationSets():
-    action = app.backend.runSearchRnaQuantificationSets
-    return handleFlaskPostRequest(flask.request, action)
-
-
-@DisplayedRoute('/rnaquantifications/search', postMethod=True)
-@requires_auth
-@oidc.require_login
-@requires_token
-def searchRnaQuantifications():
-    action = app.backend.runSearchRnaQuantifications
-    return handleFlaskPostRequest(flask.request, action)
-
-
-@DisplayedRoute('/expressionlevels/search', postMethod=True)
-@requires_auth
-@oidc.require_login
-@requires_token
-def searchExpressionLevels():
-    action = app.backend.runSearchExpressionLevels
-    return handleFlaskPostRequest(flask.request, action)
-
-
-@DisplayedRoute(
-    '/variantsets/<no(search):id>',
-    pathDisplay='/variantsets/<id>')
-@requires_auth
-@oidc.require_login
-@requires_token
-def getVariantSet(id):
-    action = app.backend.runGetVariantSet
-    return handleFlaskGetRequest(id, flask.request, action)
-
-
-@DisplayedRoute(
-    '/variants/<no(search):id>',
-    pathDisplay='/variants/<id>')
-@requires_auth
-@oidc.require_login
-@requires_token
-def getVariant(id):
-    action = app.backend.runGetVariant
-    return handleFlaskGetRequest(id, flask.request, action)
-
-
-@DisplayedRoute(
-    '/readgroupsets/<no(search):id>',
-    pathDisplay='/readgroupsets/<id>')
-@requires_auth
-@oidc.require_login
-@requires_token
-def getReadGroupSet(id):
-    action = app.backend.runGetReadGroupSet
-    return handleFlaskGetRequest(id, flask.request, action)
-
-
-@DisplayedRoute('/readgroups/<id>')
-@requires_auth
-@oidc.require_login
-@requires_token
-def getReadGroup(id):
-    action = app.backend.runGetReadGroup
-    return handleFlaskGetRequest(id, flask.request, action)
-
-
-@DisplayedRoute(
-    '/callsets/<no(search):id>',
-    pathDisplay='/callsets/<id>')
-@requires_auth
-@oidc.require_login
-@requires_token
-def getCallSet(did):
-    action = app.backend.runGetCallSet
-    return handleFlaskGetRequest(id, flask.request, action)
-
-
-@DisplayedRoute(
-    '/featuresets/<no(search):id>',
-    pathDisplay='/featuresets/<id>')
-@requires_auth
-@oidc.require_login
-@requires_token
-def getFeatureSet(id):
-    return handleFlaskGetRequest(
-        id, flask.request, app.backend.runGetFeatureSet)
-
-
-@DisplayedRoute(
-    '/features/<no(search):id>',
-    pathDisplay='/features/<id>')
-@requires_auth
-@oidc.require_login
-@requires_token
-def getFeature(id):
-    return handleFlaskGetRequest(
-        id, flask.request, app.backend.runGetFeature)
-
-
-@DisplayedRoute(
-    '/continuoussets/<no(search):id>',
-    pathDisplay='/continuoussets/<id>')
-@requires_auth
-@oidc.require_login
-@requires_token
-def getcontinuousSet(id):
-    return handleFlaskGetRequest(
-        id, flask.request, app.backend.runGetContinuousSet)
-
-
-@DisplayedRoute(
-    '/rnaquantificationsets/<no(search):id>',
-    pathDisplay='/rnaquantificationsets/<id>')
-@requires_auth
-@oidc.require_login
-@requires_token
-def getRnaQuantificationSet(id):
-    return handleFlaskGetRequest(
-        id, flask.request, app.backend.runGetRnaQuantificationSet)
-
-
-@DisplayedRoute(
-    '/rnaquantifications/<no(search):id>',
-    pathDisplay='/rnaquantifications/<id>')
-@requires_auth
-@oidc.require_login
-@requires_token
-def getRnaQuantification(id):
-    return handleFlaskGetRequest(
-        id, flask.request, app.backend.runGetRnaQuantification)
-
-
-@DisplayedRoute(
-    '/expressionlevels/<no(search):id>',
-    pathDisplay='/expressionlevels/<id>')
-@requires_auth
-@oidc.require_login
-@requires_token
-def getExpressionLevel(id):
-    return handleFlaskGetRequest(
-        id, flask.request, app.backend.runGetExpressionLevel)
-
-
-@app.route('/oauth2callback', methods=['GET'])
-def oidcCallback():
-    """
-    Once the authorization provider has cleared the user, the browser
-    is returned here with a code. This function takes that code and
-    checks it with the authorization provider to prove that it is valid,
-    and get a bit more information about the user (which we don't use).
-
-    A token is generated and given to the user, and the authorization info
-    retrieved above is stored against this token. Later, when a client
-    connects with this token, it is assumed to be a valid user.
-
-    :return: A display of the authentication token to use in the client. If
-    OIDC is not configured, raises a NotImplementedException.
-    """
-    if app.oidcClient is None:
-        raise exceptions.NotImplementedException()
-
-    response = dict(flask.request.args.iteritems(multi=True))
-    aresp = app.oidcClient.parse_response(
-        message.AuthorizationResponse,
-        info=response,
-        sformat='dict')
-    sessState = flask.session.get('state')
-    respState = aresp['state']
-    if (not isinstance(aresp, message.AuthorizationResponse) or
-            respState != sessState):
-        raise exceptions.NotAuthenticatedException()
-
-    args = {
-        "code": aresp['code'],
-        "redirect_uri": app.oidcClient.redirect_uris[0],
-        "client_id": app.oidcClient.client_id,
-        "client_secret": app.oidcClient.client_secret
-    }
-
-    atr = app.oidcClient.do_access_token_request(
-        scope="openid",
-        state=respState,
-        request_args=args)
-    if not isinstance(atr, message.AccessTokenResponse):
-        raise exceptions.NotAuthenticatedException()
-
-    atrDict = atr.to_dict()
-    if flask.session.get('nonce') != atrDict['id_token']['nonce']:
-        raise exceptions.NotAuthenticatedException()
-    key = oic.oauth2.rndstr(SECRET_KEY_LENGTH)
-    flask.session['key'] = key
-    token_data = aresp["code"], respState, atrDict
-    app.cache.set(key, token_data)
-    # flask.url_for is broken. It relies on SERVER_NAME for both name
-    # and port, and defaults to 'localhost' if not found. Therefore
-    # we need to fix the returned url
-    indexUrl = flask.url_for('index', _external=True)
-    indexParts = list(urlparse.urlparse(indexUrl))
-    if ':' not in indexParts[1]:
-        indexParts[1] = '{}:{}'.format(socket.gethostname(), app.myPort)
-        indexUrl = urlparse.urlunparse(indexParts)
-    response = flask.redirect(indexUrl)
-    return response
-
-
-# Leaving this function here for reference for the Authorization Code flow
-# Added by Kevin Chan
-@app.route('/keycallback')
-def keycloakCallback():
-    """
-    Similar to the oidcCallback function, once the authorization provider
-    has cleared the user, browser is returned here with a code. The code
-    is then checked with the authorization provider and if valid a token
-    is returned.
-
-    The token is stored in the session, and the user is assumed to be valid.
-
-    Returns: a token and the redirect url to the new page.
-    """
-    if app.oidcClient is None:
-        raise exceptions.NotImplementedException()
-
-    response = dict(flask.request.args.iteritems(multi=True))
-    authMsg = message.AuthorizationResponse
-    aresp = app.oidcClient.parse_response(
-        authMsg, info=response, sformat="dict")
-    respState = aresp["state"]
-    sessState = flask.session.get('state')
-
-    if sessState != respState:
-        raise exceptions.NotAuthorizedException()
-
-    args = {
-        "code": aresp["code"],
-        "redirect_uri": app.oidcClient.redirect_uris[0],
-        "client_id": app.oidcClient.client_id,
-        "client_secret": app.oidcClient.client_secret,
-        "token_endpoint": app.oidcClient.provider_info["token_endpoint"],
-        "grant_type": "authorization_code",
-    }
-
-    endpoint = app.oidcClient.provider_info["token_endpoint"]
-    tokResp = requests.post(url=endpoint, data=args)
-    tokContent = json.loads(tokResp.content)
-    # token = tokContent["access_token"]
-
-    if tokResp.status_code != 200:
-        raise exceptions.NotAuthorizedException()
-    app.oidcClient.token = tokContent
-
-    # flask.session["key"] = idGenerator(size=SECRET_KEY_LENGTH)
-    # This next line will display the access token on the server front end.
-    # If you do not want this uncomment the line above
-    flask.session["key"] = app.oidcClient.token["access_token"]
-    # change the url depending on where the GA4GH server is hosted
-    hostname = socket.gethostname()
-    path = flask.session["path"]
-    host = socket.gethostbyname(hostname, app.myPort, path)
-    redirectUri = 'http://{0}:{1}{2}'.format(host)
-    return flask.redirect(redirectUri)
-
-
-@DisplayedRoute(
-    '/datasets/<no(search):id>',
-    pathDisplay='/datasets/<id>')
-@requires_auth
-@oidc.require_login
-@requires_token
-def getDataset(id):
-    return handleFlaskGetRequest(
-        id, flask.request, app.backend.runGetDataset)
-
-
-@DisplayedRoute(
-    '/variantannotationsets/<no(search):id>',
-    pathDisplay='/variantannotationsets/<id>')
-@requires_auth
-@oidc.require_login
-@requires_token
-def getVariantAnnotationSet(id):
-    return handleFlaskGetRequest(
-        id, flask.request, app.backend.runGetVariantAnnotationSet)
-
-
-@DisplayedRoute('/phenotypes/search', postMethod=True)
-@requires_auth
-@oidc.require_login
-@requires_token
-def searchPhenotypes():
-    return handleFlaskPostRequest(
-        flask.request, app.backend.runSearchPhenotypes)
-
-
-@DisplayedRoute('/featurephenotypeassociations/search', postMethod=True)
-@requires_auth
-@oidc.require_login
-@requires_token
-def searchGenotypePhenotypes():
-    return handleFlaskPostRequest(
-        flask.request,
-        app.backend.runSearchGenotypePhenotypes)
-
-
-@DisplayedRoute('/phenotypeassociationsets/search', postMethod=True)
-@requires_auth
-@oidc.require_login
-@requires_token
-def searchPhenotypeAssociationSets():
-    return handleFlaskPostRequest(
-        flask.request, app.backend.runSearchPhenotypeAssociationSets)
-
-
-# The below methods ensure that JSON is returned for various errors
-# instead of the default, html
-
-@app.errorhandler(401)
-def unauthorizedHandler(errorString):
-    return handleException(exceptions.UnauthorizedException(errorString))
-
-
-@app.errorhandler(404)
-def pathNotFoundHandler(errorString):
-    return handleException(exceptions.PathNotFoundException())
-
-
-@app.errorhandler(405)
-def methodNotAllowedHandler(errorString):
-    return handleException(exceptions.MethodNotAllowedException())
-
-
-@app.errorhandler(403)
-def notAuthenticatedHandler(errorString):
-    return handleException(exceptions.NotAuthenticatedException(errorString))
