+++ ga4gh/server/frontend/static/favicon.ico
