+++ ga4gh/server/frontend/templates/initial_peers.txt
